+++ projects/kypo-training-model/src/lib/cheating-detection/detection-events/detected-forbidden-command.ts
@@ -1,0 +1,12 @@
+/**
+ * This class represents a forbidden command
+ * used in a forbidden command detection event
+ */
+import { DetectedForbiddenCommandTypeEnum } from '../../enums/detected-forbidden-command-type.enum';
+
+export class DetectedForbiddenCommand {
+  command: string;
+  type: DetectedForbiddenCommandTypeEnum;
+  hostname: string;
+  occurredAt: Date;
+}