+++ projects/kypo-training-model/src/lib/cheating-detection/detection-events/detection-event.ts
@@ -5,11 +5,14 @@ import { AbstractDetectionEventTypeEnum } from '../../enums/abstract-detection-e
 
 export class AbstractDetectionEvent {
   id: number;
+  trainingRunId: number;
   trainingInstanceId: number;
   cheatingDetectionId: number;
   detectionEventType: AbstractDetectionEventTypeEnum;
   detectedAt: Date;
   levelTitle: string;
   levelId: number;
+  levelOrder: number;
   participantCount: number;
+  participants: string;
 }