+++ projects/kypo-training-model/src/lib/cheating-detection/detection-events/forbidden-commands-detection-event.ts
@@ -1,10 +1,11 @@
 import { AbstractDetectionEvent } from './detection-event';
+import { ForbiddenCommand } from './forbidden-command';
 
 /**
  * Class representing detection event of type Forbidden Commands
  */
 export class ForbiddenCommandsDetectionEvent extends AbstractDetectionEvent {
-  forbiddenCommands: string[];
+  commandCount: number;
   constructor() {
     super();
   }